chore(wtr): move WTR plugins from helpers dir to config dir
